server/requirements.txt
annotated-types==0.7.0
anyio<4
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.0.0
flask-cors==4.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.3.7
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
PyPDF2==3.0.1
python-docx==1.0.1
